Refactor: merged all tests regarding FieldUtils into one file
